mm/mempolicy: modify interleave mempolicy to use memtier weights
The memory-tier subsystem implements interleave weighting
for tiers for the purpose of bandwidth optimization.  Each
tier may contain multiple numa nodes, and each tier may have
different weights in relation to each compute node.

By default, all memory tiers have a weight of 1, which retains
the default mempolicy interleave behavior of an even distribution.

When a tier weight is changed, nodes in that tier will inherit
the weight when distributing memory with an interleave policy.

Weights are relative to the node requesting it (i.e. the weight
for tier2 from node0 may be different than the weight for tier2
from node1).  This allows for cpu-bound tasks to have more
precise control over the distribution of memory.

For example consider 2 tiers:
* tier1 nodes: [0,1] - weights [4,3]
* tier2 nodes: [2,3] - weights [2,1]

Task A:
   cpunode: 0
   nodemask: 0,2
   allocation result:  [0,0,0,0,2,2, repeat]

Task B:
   cpunode: 1
   nodemask: 1,3
   allocation result:  [1,1,1,3, repeat]

Task C:
   cpunode: 0
   nodemask: 0,1,2,3
   allocation result:  [0,0,0,0,1,1,1,1,2,2,3,3, repeat]

Task D:
   cpunode: 1
   nodemask: 0,1,2,3
   allocation result:  [0,0,0,1,1,1,2,3, repeat]

The minimum weight for a valid node is always 1.

In the future, it is expected that nodes may be shifted between tiers
and new tiers created to allow for even more relational weightings.

In particular, as of this patch all DRAM nodes are located in a single
tier, which may create a sub-optimal distribution if there exists a
bandwidth chokepoint between sockets.

Signed-off-by: Gregory Price <gregory.price@memverge.com>