mm/mempolicy: add node weights to MPOL_INTERLEAVE
By default, interleave distributes memory in a simple round-robin fashion,
but NUMA nodes can have different latency and bandwidth attributes, making
the optimal distribution of memory for a given task non-uniform.

Add an array of 'char weights' to mempolicy, and add an fs/proc entry
which allows software to write weights to the mempolicy

echo "0:10,1:1" > /proc/pid/interleave_weights

For each interleave round, this will cause 10 pages to be allocated on
node 0 and 1 page to be allocated on node 1.

By default, all nodes defined in nodemask_t when MPOL_INTERLEAVE is set
will have the node weight set to 1. This retains the original behavior
of MPOL_INTERLEAVE.

Weights can not be set to 0, and should a weight of 0 be detected while
attempting to interleave nodes, instead 1 allocation will be made. If a
node weight of 0 is desired, the nodemask should be changed to omit that
node from the interleave set.

When mpol weights are updated from fs/proc, the current mempolicy for the
task is duplicated, updated, and replaced with the task lock held. This
avoids any race conditions any active interleave functions and any cgroup
rebindings that may be occurring at the same time.

Signed-off-by: Gregory Price <gregory.price@memverge.com>