node: add interleave weights to node hmem information
When hmem information is available, a configurable interleave
weight should also be exposed.  The intent of this weight is to
enable set_mempolicy(MPOL_INTERLEAVE) to distribute memory across
nodes based on the accessor and the effective bandwidth available.

The end goal is to maximize the effect use of total bandwidth.

Signed-off-by: Gregory Price <gregory.price@memverge.com>