node: add interleave weights to node accessor
Add a configurable interleave weight to the node for each possible
accessor. The intent of this weight is to enable set_mempolicy() to
to distribute memory across nodes based on the accessor and the
effective bandwidth available.

The goal is to maximize the effective use of available bandwidth.

The default weight is 1 for all nodes, which will mimic the current
interleave (basic round-robin).

Signed-off-by: Gregory Price <gregory.price@memverge.com>
Suggested-by: Ying Huang <ying.huang@intel.com>
Suggested-by: Ravi Shankar <ravis.opensrc@micron.com>