mm/mempolicy: modify do_mbind to operate on task/mm struct argst
To make mbind applied mempolicies modifable by external tasks, we must
first change the do_mbind callstack to take a task and its mm as an
argument.  It is expected that the caller will hold references to the
respective task and mm.

The task is required for calls to get_task_vma_policy and
get_task_policy, while the mm is required for access to vma policies.
The mm argument was added to generalize the code paths between the
original mbind syscall and the upcoming pidfd_syscall, as the original
syscall operates directly on (current, current->mm), while the new
syscall will require taking references to each.

get_vma_policy is changed into a wrapper of get_task_vma_policy which
passes current as an argument to retain the existing behavior for
callers of get_vma_policy.

Since task->policy is no longer guaranteed to be stable, any time we
seek to acquire a policy via get_task_vma_policy, we must use the
task_lock and reference it accordingly, regardless of where it
ultimately came from.  do_get_mempolicy has a simliar behavior, where
where a reference is taken and a conditional release is made to
handle the situation where a shared policy is acquired.

In the case of do_mbind, we don't actually need to take a reference
to the policy, as we only call get_task_vma_policy to find the ilx.
In this case, we only need to call mpol_cond_put immediately to ensure
that if get_task_vma_policy returns a shared policy we decrement the
reference count since a shared mpol will return already referenced.

However, we must still acquire the task_lock to safely access
task->policy on the call to get_task_vma_policy.

Signed-off-by: Gregory Price <gregory.price@memverge.com>