mm/mempolicy: refactor interfaces to use task instead of current
In an effort to make mempolicies externally modifiable, we need
to change some internal interfaces to allow the use of non-current
tasks.  The first step is to change interfaces to accept an explicit
task instead of using current directly, and write wrappers for those
functions for existing users.

This does require taking the task lock and referencing task->mm
in a variety of places.  In some cases, the task->mm was directly
referenced due to the implication of the task being the modifier
of the task policy or accessor of the task->mm.  In this scenarios
we check if (task == current) to retain the original behavior.

This is done to avoid a check in get_task_mm that was not present
in the original path:

  if (mm) {
    if (task->flags & PF_KTHREAD)
      mm = NULL;
    else
      mmget(mm);
  }

When (task == curent), if we use get_task_mm, it would prevent a
kernel task from making modifications or accessing information
about its own vma's.  So in this scenario, we simply access and
reference the mm directly, since the mempolicy information is
being accessed in the context of the task itself.

The retains the existing behavior.

Signed-off-by: Gregory Price <gregory.price@memverge.com>