mm/mempolicy: mpol_parse_str should ignore trailing characters in nodelist
When validating MPOL_PREFERRED, the nodelist has already been parsed
and error checked by nodelist_parse.  So rather than looping through
the string again, we should just check that the weight of the nodemask
is 1, which is the actual condition we care to check.

This also handles the case where newline characters are present.

Signed-off-by: Gregory Price <gregory.price@memverge.com>